move avatar output (i.e., configurable machine) behind log flag

+++ vl.c
@@ -4470,7 +4470,7 @@ int main_aux(int argc, char **argv, char **envp, PandaMainMode pmm)
     }
 
     if (log_mask) {
-        int mask;
+        long mask;
         mask = qemu_str_to_log_mask(log_mask);
         if (!mask) {
             qemu_print_log_usage(stdout);
